[IMP] stock_split_picking: Add a common for tests
